[frontend] no results if no text keyword (#7279)
